cfa

compute counterfactuals using distribution regression with a continuous treatment

cfa: A package for computing counterfactuals with a continuous treatment

cfa(formla, xformla = NULL, tvals, yvals, data, method = "dr",
  link = "logit", tau = seq(0.01, 0.99, 0.01), condDistobj = NULL,
  se = TRUE, iters = 100, cl = 1)

Arguments

formla
a formula y ~ treatment

xformla
one sided formula for x variables to include, e.g. ~x1 + x2

tvals
the values of the "treatment" to compute parameters of interest for

yvals
the values to compute the counterfactual distribution for

data
the data.frame where y, t, and x are

method
either "dr" or "qr" for distribution regression or quantile regression

link
if using distribution regression, any link function that works with the binomial family (e.g. logit (the default), probit, cloglog)

tau
if using quantile regression, which values of tau to estimate the conditional quantiles

condDistobj
optional conditional distribution object that has been previously computed

se
whether or not to compute standard errors using the bootstrap

iters
how many bootstrap iterations to use

cl
how many clusters to use for parallel computation of standard errors

Value

CFA object

Examples

data(igm)
tvals <- seq(10,12,length.out=10)
yvals <- seq(quantile(igm$lcfincome, .05), quantile(igm$lcfincome, .95), length.out=50)
## This line doesn't adjust for any covariates
cfa(lcfincome ~ lfincome, tvals=tvals, yvals=yvals, data=igm,
 se=FALSE)

#> Error in `[.data.frame`(data, , yname): undefined columns selected

## This line adjusts for differences in education
cfa(lcfincome ~ lfincome, ~HEDUC, tvals=tvals, yvals=yvals, data=igm,
 se=FALSE)

#> Error in `[.data.frame`(data, , yname): undefined columns selected